requirements.txt
numpy~=2.0.0
torch~=2.3.1
torchvision~=0.18.1
pillow~=10.3.0
opencv-python~=4.10.0.84
